Implement profiler
Apllication can be started with cProfile from command window. The
command is 'python main.py profiler [time in seconds]'. This way the
application starts and a timer stops it after the given time. If zero is
given as time, the application will stop when the user closes it.
Yielded profile statistics are saved in the working directory. The
filename contains the timestamp of profiling.