Find cause of software lag
Lag in treadmill state visualization and feedback missmatches are caused
by serial buffer filling up when the software is used for prolonged
time. Most probably the accumulation of data packages is caused by
inconsistent timeout of sampling QTimer which is programmed to shot in
every 5 ms. The average timeout is indeed around 5 ms, but sometimes it can be over 10 ms. A solution could be that in every reading cycle the software reads out all data packages in the buffer, even if it's more than one. For some reason, this way the recorded data has replicated timestamps and also missing ones. Next step is to resolve this issue.