Overcome data record bug
Data record bug (mentioned in previous commit) was caused by improper
data logging. The variable that was written into the file was not
updated in every data assignment but in every data read timer timeout
(sampling timer). Now the mechanism is revired by using signals and data
is recorded properly. However, refactorisation is needed for a cleaner
code.